refactor(ios): subview decomposition + Helpers/ + strict lint (PUL-209)
Phase 5 of BudgetDetails refactor. Splits oversized files (>350 LOC),
extracts the three Task.sleep call sites into dedicated Helpers/ modifiers,
and escalates `no_task_sleep_in_feature` + `no_magic_timing_literals` from
warning to error. Architecture tests for both invariants now run strict.

- Helpers/AutoPopView.swift, View+afterPushTransition.swift,
  View+syncIndicatorRamp.swift centralise the only legitimate Task.sleep
  uses (autopop grace, autofocus push delay, sync indicator ramp).
- BudgetLineDetailPage 427 → 277 LOC (Hero / Title and DEBUG harness moved
  to BudgetLineDetailPage+Hero.swift / +VerifyHarness.swift).
- BudgetLineMixedRow 484 → 303 LOC (PointCircle promoted to its own file,
  previews moved to BudgetLineMixedRow+Previews.swift).
- BudgetDetailHero 389 → 331 LOC (previews moved to +Previews.swift).
- AddAllocatedTransactionPage / EditTransactionPage / BudgetLineMixedRow
  drop inline Task.sleep + magic ms literals; consume the shared helpers.
- 11 architecture tests now strict (no `withKnownIssue` wrappers).
