+++ ios/Pulpe/Features/Budgets/BudgetDetails/BudgetLineMixedRow.swift
@@ -301,184 +301,3 @@ struct BudgetLineMixedRow: View {
         return "\(kindWord) · \(line.name) · \(amount) · \(pointed)"
     }
 }
-
-// MARK: - PointCircle (iOS Reminders-style toggle)
-
-/// 44×44 hit area, 24pt visible circle. Tap toggles via `onToggle` callback;
-/// the parent owns state and animates `isPointed` flips. Uses `Button` so the
-/// tap is visible to VoiceOver and can be hit independently of the parent row's
-/// own tap target. Apple HIG: 44pt minimum tap area.
-private struct PointCircle: View {
-    let isPointed: Bool
-    let color: Color
-    let isSyncing: Bool
-    let onToggle: () -> Void
-
-    /// Debounced sync state — only flips true if the toggle takes >300 ms,
-    /// so fast optimistic updates don't trigger a green-dot flash.
-    @State private var displayedSyncing = false
-
-    var body: some View {
-        Button(action: onToggle) {
-            ZStack {
-                Circle()
-                    .fill(isPointed ? color : Color.clear)
-                    .overlay {
-                        Circle()
-                            .strokeBorder(
-                                isPointed ? color : Color.outlineVariant,
-                                lineWidth: DesignTokens.BorderWidth.thick
-                            )
-                    }
-                    .frame(
-                        width: DesignTokens.Checkbox.size,
-                        height: DesignTokens.Checkbox.size
-                    )
-
-                if isPointed {
-                    Image(systemName: "checkmark")
-                        .font(.caption2.weight(.bold))
-                        .foregroundStyle(.white)
-                        .transition(.scale.combined(with: .opacity))
-                }
-
-                if displayedSyncing {
-                    SyncIndicator(isSyncing: true)
-                        .offset(x: DesignTokens.Checkbox.size / 2 - 2, y: -DesignTokens.Checkbox.size / 2 + 2)
-                }
-            }
-            .frame(
-                width: DesignTokens.TapTarget.minimum,
-                height: DesignTokens.TapTarget.minimum
-            )
-            .contentShape(Circle())
-        }
-        .buttonStyle(.plain)
-        .task(id: isSyncing) {
-            guard isSyncing else {
-                displayedSyncing = false
-                return
-            }
-            try? await Task.sleep(for: .milliseconds(300))
-            guard !Task.isCancelled else { return }
-            displayedSyncing = true
-        }
-        .accessibilityLabel(isPointed ? "Pointé" : "À pointer")
-        .accessibilityAddTraits(isPointed ? [.isButton, .isSelected] : .isButton)
-    }
-}
-
-// MARK: - Previews
-
-private extension BudgetLine {
-    static func preview(
-        id: String = UUID().uuidString,
-        name: String,
-        amount: Decimal,
-        kind: TransactionKind,
-        recurrence: TransactionRecurrence = .fixed,
-        isChecked: Bool = false
-    ) -> BudgetLine {
-        BudgetLine(
-            id: id,
-            budgetId: "preview-budget",
-            templateLineId: nil,
-            savingsGoalId: nil,
-            name: name,
-            amount: amount,
-            kind: kind,
-            recurrence: recurrence,
-            isManuallyAdjusted: false,
-            checkedAt: isChecked ? Date() : nil,
-            createdAt: Date(),
-            updatedAt: Date()
-        )
-    }
-}
-
-private struct BudgetLineMixedRowPreviewHost: View {
-    let cases: [(line: BudgetLine, consumption: BudgetFormulas.Consumption)]
-
-    var body: some View {
-        ScrollView {
-            VStack(spacing: DesignTokens.Spacing.sm) {
-                ForEach(Array(cases.enumerated()), id: \.offset) { _, item in
-                    BudgetLineMixedRow(
-                        line: item.line,
-                        consumption: item.consumption,
-                        isSyncing: false,
-                        currency: .chf,
-                        onTap: {},
-                        onTogglePointed: {}
-                    )
-                }
-            }
-            .padding(DesignTokens.Spacing.lg)
-        }
-        .background(Color.appBackground)
-    }
-}
-
-#Preview("Expense — empty (no real)") {
-    let line = BudgetLine.preview(name: "Téléphone", amount: 100, kind: .expense)
-    // Literal zero consumption — projector pre-computes this at runtime; previews
-    // mock it so the file stays compliant with `no_formula_in_view_body`.
-    let consumption = BudgetFormulas.Consumption(allocated: 0, available: line.amount, percentage: 0)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Expense — partial (260 CHF restant)") {
-    let line = BudgetLine.preview(name: "Courses", amount: 800, kind: .expense)
-    let consumption = BudgetFormulas.Consumption(allocated: 540, available: 260, percentage: 67.5)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Expense — over budget") {
-    let line = BudgetLine.preview(name: "Sorties & loisirs", amount: 300, kind: .expense)
-    let consumption = BudgetFormulas.Consumption(allocated: 320, available: -20, percentage: 106.7)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Income — partially received") {
-    let line = BudgetLine.preview(name: "Salaire", amount: 7500, kind: .income)
-    let consumption = BudgetFormulas.Consumption(allocated: 3000, available: 4500, percentage: 40)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Saving — fully transferred") {
-    let line = BudgetLine.preview(name: "Épargne du mois", amount: 600, kind: .saving)
-    let consumption = BudgetFormulas.Consumption(allocated: 600, available: 0, percentage: 100)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Pointed (dimmed)") {
-    let line = BudgetLine.preview(name: "Loyer", amount: 2100, kind: .expense, isChecked: true)
-    let consumption = BudgetFormulas.Consumption(allocated: 2100, available: 0, percentage: 100)
-    return BudgetLineMixedRowPreviewHost(cases: [(line, consumption)])
-}
-
-#Preview("Mixed list") {
-    let income = BudgetLine.preview(name: "Salaire", amount: 7500, kind: .income)
-    let incomeConsumption = BudgetFormulas.Consumption(allocated: 7500, available: 0, percentage: 100)
-
-    let saving = BudgetLine.preview(name: "Épargne du mois", amount: 600, kind: .saving)
-    let savingConsumption = BudgetFormulas.Consumption(allocated: 600, available: 0, percentage: 100)
-
-    let phone = BudgetLine.preview(name: "Téléphone", amount: 100, kind: .expense)
-    // Literal zero consumption — projector pre-computes at runtime; previews mock it.
-    let phoneConsumption = BudgetFormulas.Consumption(allocated: 0, available: phone.amount, percentage: 0)
-
-    let groceries = BudgetLine.preview(name: "Courses", amount: 800, kind: .expense)
-    let groceriesConsumption = BudgetFormulas.Consumption(allocated: 540, available: 260, percentage: 67.5)
-
-    let leisure = BudgetLine.preview(name: "Sorties & loisirs", amount: 300, kind: .expense)
-    let leisureConsumption = BudgetFormulas.Consumption(allocated: 320, available: -20, percentage: 106.7)
-
-    return BudgetLineMixedRowPreviewHost(cases: [
-        (income, incomeConsumption),
-        (saving, savingConsumption),
-        (phone, phoneConsumption),
-        (groceries, groceriesConsumption),
-        (leisure, leisureConsumption),
-    ])
-}
